+++ src/Paket.Bootstrapper/Properties/AssemblyInfo.cs
@@ -4,11 +4,12 @@ using System.Reflection;
 [assembly: AssemblyTitleAttribute("Paket.Bootstrapper")]
 [assembly: AssemblyProductAttribute("Paket")]
 [assembly: AssemblyDescriptionAttribute("A package dependency manager for .NET with support for NuGet packages and GitHub repositories.")]
-[assembly: AssemblyVersionAttribute("2.57.1")]
-[assembly: AssemblyFileVersionAttribute("2.57.1")]
-[assembly: AssemblyInformationalVersionAttribute("2.57.1")]
+[assembly: AssemblyVersionAttribute("2.58.7")]
+[assembly: AssemblyFileVersionAttribute("2.58.7")]
+[assembly: AssemblyInformationalVersionAttribute("2.58.7")]
 namespace System {
     internal static class AssemblyVersionInformation {
-        internal const string Version = "2.57.1";
+        internal const string Version = "2.58.7";
+        internal const string InformationalVersion = "2.58.7";
     }
 }