+++ src/Paket.PowerShell/AssemblyInfo.fs
@@ -5,10 +5,11 @@ open System.Reflection
 [<assembly: AssemblyProductAttribute("Paket")>]
 [<assembly: AssemblyCompanyAttribute("Paket team")>]
 [<assembly: AssemblyDescriptionAttribute("A package dependency manager for .NET with support for NuGet packages and GitHub repositories.")>]
-[<assembly: AssemblyVersionAttribute("2.57.1")>]
-[<assembly: AssemblyFileVersionAttribute("2.57.1")>]
-[<assembly: AssemblyInformationalVersionAttribute("2.57.1")>]
+[<assembly: AssemblyVersionAttribute("2.58.7")>]
+[<assembly: AssemblyFileVersionAttribute("2.58.7")>]
+[<assembly: AssemblyInformationalVersionAttribute("2.58.7")>]
 do ()
 
 module internal AssemblyVersionInformation =
-    let [<Literal>] Version = "2.57.1"
+    let [<Literal>] Version = "2.58.7"
+    let [<Literal>] InformationalVersion = "2.58.7"