+++ RELEASE_NOTES.md
@@ -1,4 +1,23 @@
-#### 2.57.1 - 30.03.2016
+#### 2.58.7 - 07.04.2016
+* Paket update supports combining filter with specific version - https://github.com/fsprojects/Paket/pull/1580
+
+#### 2.58.6 - 07.04.2016
+* BUGFIX: Redirects for assemblies in the GAC were removed - https://github.com/fsprojects/Paket/issues/1574
+
+#### 2.58.5 - 07.04.2016
+* BUGFIX: Paket.dependency with version ranges failed when package has pinned dependency and that version is unlisted - https://github.com/fsprojects/Paket/issues/1579
+
+#### 2.58.4 - 07.04.2016
+* BUGFIX: Github dependencies reference transitive NuGet packages to projects - https://github.com/fsprojects/Paket/issues/1578
+
+#### 2.58.3 - 05.04.2016
+* BUGFIX: Add "*.fsi" files as <Compile> by default - https://github.com/fsprojects/Paket/pull/1573
+
+#### 2.58.2 - 05.04.2016
+* Touch: disabled by default in Add, Update, Install; enabled with --touch-affected-refs - https://github.com/fsprojects/Paket/pull/1571
+
+#### 2.58.1 - 05.04.2016
+* Using saved api key credentials for the push operation - https://github.com/fsprojects/Paket/pull/1570
 * BUGFIX: Property Definitions: placed after csharp targets - https://github.com/fsprojects/Paket/pull/1522
 
 #### 2.57.0 - 30.03.2016